Sletter sparkel-stubs
