+++ src/main/kotlin/no/nav/helse/App.kt
@@ -20,11 +20,7 @@ import io.ktor.util.KtorExperimentalAPI
 import io.prometheus.client.hotspot.DefaultExports
 import no.nav.helse.aktoer.AktoerGateway
 import no.nav.helse.aktoer.AktoerService
-import no.nav.helse.arbeidsgiver.ArbeidsgiverGateway
-import no.nav.helse.arbeidsgiver.ArbeidsgiverService
 import no.nav.helse.arbeidsgiver.arbeidsgiverApis
-import no.nav.helse.barn.BarnGateway
-import no.nav.helse.barn.BarnService
 import no.nav.helse.barn.barnApis
 import no.nav.helse.dusseldorf.ktor.auth.clients
 import no.nav.helse.dusseldorf.ktor.client.HttpRequestHealthCheck
@@ -38,19 +34,20 @@ import no.nav.helse.dusseldorf.ktor.jackson.JacksonStatusPages
 import no.nav.helse.dusseldorf.ktor.jackson.dusseldorfConfigured
 import no.nav.helse.dusseldorf.ktor.metrics.MetricsRoute
 import no.nav.helse.dusseldorf.ktor.metrics.init
-import no.nav.helse.general.auth.*
+import no.nav.helse.general.auth.IdTokenProvider
+import no.nav.helse.general.auth.authorizationStatusPages
 import no.nav.helse.general.systemauth.AccessTokenClientResolver
 import no.nav.helse.k9.K9OppslagArbeidsgivereService
 import no.nav.helse.k9.K9OppslagBarnService
 import no.nav.helse.k9.K9OppslagGateway
 import no.nav.helse.k9.K9OppslagSokerService
-import no.nav.helse.person.PersonGateway
-import no.nav.helse.person.PersonService
 import no.nav.helse.soker.sokerApis
 import no.nav.helse.soknad.PleiepengesoknadMottakGateway
 import no.nav.helse.soknad.SoknadService
 import no.nav.helse.soknad.soknadApis
-import no.nav.helse.vedlegg.*
+import no.nav.helse.vedlegg.PleiepengerDokumentGateway
+import no.nav.helse.vedlegg.VedleggService
+import no.nav.helse.vedlegg.vedleggApis
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import java.time.Duration
@@ -139,26 +136,6 @@ fun Application.pleiepengesoknadapi() {
             )
         )
 
-        val personGateway = PersonGateway(
-            baseUrl = configuration.getSparkelUrl(),
-            accessTokenClient = accessTokenClientResolver.sparkel(),
-            apiGatewayApiKey = apiGatewayApiKey
-        )
-
-        val barnGateway = BarnGateway(
-            baseUrl = configuration.getSparkelUrl(),
-            aktoerService = aktoerService,
-            accessTokenClient = accessTokenClientResolver.sparkel(),
-            apiGatewayApiKey = apiGatewayApiKey
-        )
-
-        val arbeidsgiverGateway = ArbeidsgiverGateway(
-            aktoerService = aktoerService,
-            baseUrl = configuration.getSparkelUrl(),
-            accessTokenClient = accessTokenClientResolver.sparkel(),
-            apiGatewayApiKey = apiGatewayApiKey
-        )
-
         val pleiepengesoknadMottakGateway = PleiepengesoknadMottakGateway(
             baseUrl = configuration.getPleiepengesoknadMottakBaseUrl(),
             accessTokenClient = accessTokenClientResolver.pleiepengesoknadMottak(),
@@ -213,9 +190,7 @@ fun Application.pleiepengesoknadapi() {
         val healthService = HealthService(
             healthChecks = setOf(
                 aktoerGateway,
-                personGateway,
-                barnGateway,
-                arbeidsgiverGateway,
+                k9OppslagGateway,
                 pleiepengesoknadMottakGateway,
                 HttpRequestHealthCheck(mapOf(
                     configuration.getJwksUrl() to HttpRequestHealthConfig(expectedStatus = HttpStatusCode.OK, includeExpectedStatusEntity = false),
