+++ src/test/kotlin/no/nav/helse/ApplicationWithMocks.kt
@@ -32,9 +32,6 @@ class ApplicationWithMocks {
                 .stubAktoerRegisterGetAktoerId()
                 .stubLeggSoknadTilProsessering()
                 .stubPleiepengerDokument()
-                .stubSparkelGetArbeidsgivere()
-                .stubSparkelGetBarn()
-                .stubSparkelGetPerson()
                 .stubK9OppslagSoker()
 
             val testArgs = TestConfiguration.asMap(
